Add adapter class
