+++ app/src/main/java/com/example/cloneinstagram/main/home/DetailViewFragment.kt
@@ -31,124 +31,11 @@ class DetailViewFragment : Fragment() {
         firestore = FirebaseFirestore.getInstance()
         uid = FirebaseAuth.getInstance().currentUser?.uid
 
-        view.detailviewfragment_recyclerview.adapter = DetailViewRecyclerViewAdapter()
+        view.detailviewfragment_recyclerview.adapter = DetailViewRecyclerAdapter()
         view.detailviewfragment_recyclerview.layoutManager = LinearLayoutManager(activity)
         return view
     }
 
-    inner class DetailViewRecyclerViewAdapter : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
-        var contentDTOs: ArrayList<ContentDTO> = arrayListOf()
-        var contentUidList: ArrayList<String> = arrayListOf()
 
-        init {
-           firestore?.collection("images")?.orderBy("timestamp")?.addSnapshotListener { querySnapshot, firebaseFirestoreException ->
-               contentDTOs.clear()
-               contentUidList.clear()
-                //Sometimes, this code return null of queryshanpshot when it signout
-               if(querySnapshot == null) return@addSnapshotListener
-
-               for(snapshot in querySnapshot!!.documents){
-                   var item = snapshot.toObject(ContentDTO::class.java)
-                   contentDTOs.add(item!!)
-                   contentUidList.add(snapshot.id)
-               }
-               notifyDataSetChanged()
-           }
-        }
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
-            var view = LayoutInflater.from(parent.context).inflate(R.layout.item_detail,parent,false)
-            return CustomViewHolder(view)
         }
 
-        inner class CustomViewHolder(view: View) : RecyclerView.ViewHolder(view)
-
-        override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
-            var viewholder = (holder as CustomViewHolder).itemView
-
-            //userid
-            viewholder.detailviewitem_profile_textview.text = contentDTOs!![position].userId
-
-            //Image
-            Glide.with(holder.itemView.context).load(contentDTOs!![position].imageUrl).into(viewholder.detailviewitem_imageview_content)
-
-            //Explain of content
-            viewholder.detailviewitem_explain_textview.text =contentDTOs!![position].explain
-
-            //likes
-            viewholder.detailviewitem_favoritecounter_textview.text = "Likes" + contentDTOs!![position].favoriteCount
-
-            //profileImage
-            Glide.with(holder.itemView.context).load(contentDTOs!![position].imageUrl).into(viewholder.detailviewitem_profile_image)
-
-            //when the button is clicked
-            viewholder.detailviewitem_favorite_imageview.setOnClickListener {
-                favoriteEvent(position)
-            }
-            // when the page is loaded
-            if(contentDTOs!![position].favorites.containsKey(uid)){
-                //like status
-                viewholder.detailviewitem_favorite_imageview.setImageResource(R.drawable.ic_favorite)
-                } else{
-                    //unlike status
-                    viewholder.detailviewitem_favorite_imageview.setImageResource(R.drawable.ic_favorite_border)
-                }
-            viewholder.detailviewitem_profile_image.setOnClickListener {
-                var fragment = UserFragment()
-                var bundle = Bundle()
-                bundle.putString("destinationUid",contentDTOs[position].uid)
-                bundle.putString("userId",contentDTOs[position].userId)
-                fragment.arguments = bundle
-                activity?.supportFragmentManager?.beginTransaction()?.replace(R.id.main_content,fragment)?.commit()
-
-            }
-
-            viewholder.detailviewitem_comment_imageview.setOnClickListener { v->
-                var intent = Intent(v.context,CommentActivity::class.java)
-                intent.putExtra("contentUid",contentUidList[position])
-                intent.putExtra("destinationUid",contentDTOs[position].uid)
-                startActivity(intent)
-
-            }
-            }
-
-        override fun getItemCount(): Int {
-            return contentDTOs.size
-        }
-        fun favoriteEvent(position : Int) {
-            var tsDoc = firestore?.collection("images")?.document(contentUidList[position])
-            firestore?.runTransaction{ transaction->
-
-                var uid = FirebaseAuth.getInstance().currentUser?.uid
-                var contentDTO = transaction.get(tsDoc!!).toObject(ContentDTO::class.java)
-
-                if(contentDTO!!.favorites.containsKey(uid)){
-                        //when the button is clicked
-                        contentDTO?.favoriteCount = contentDTO?.favoriteCount -1
-                        contentDTO?.favorites.remove(uid)
-                    }else{
-                        //when the button is not clicked
-                        contentDTO?.favoriteCount = contentDTO?.favoriteCount + 1
-                        contentDTO?.favorites[uid!!] = true
-                    favoriteAlarm(contentDTOs[position].uid!!)
-
-                    }
-                transaction.set(tsDoc,contentDTO)
-
-
-            }
-        }
-
-        fun favoriteAlarm(destinationUid : String) {
-            var alarmDTO = AlarmDTO()
-            alarmDTO.destinationUid = destinationUid
-            alarmDTO.userId = FirebaseAuth.getInstance().currentUser?.email
-            alarmDTO.uid = FirebaseAuth.getInstance().currentUser?.uid
-            alarmDTO.kind = 0
-            alarmDTO.timestamp = System.currentTimeMillis()
-            FirebaseFirestore.getInstance().collection("alarms").document().set(alarmDTO)
-        }
-
-    }
-
-
-}
